refactor: extract game_data reader
