+++ tmou-mapa-backend/sample_game_data.yaml
@@ -1,0 +1,166 @@
+game:
+  items:
+    item:
+      - node:
+          id: '539563487'
+        name: puzzles-1
+        type: puzzles
+        description: šifry sady 1
+        url: /static/puzzles/7-mysi_a_syr.pdf
+        level: '1'
+      - node:
+          id: '539563487'
+        name: puzzles-2
+        type: puzzles
+        description: šifry sady 2
+        url: /static/puzzles/7-mysi_a_syr.pdf
+        level: '2'
+      - node:
+          id: '539563487'
+        name: puzzles-3
+        type: puzzles
+        description: šifry sady 2
+        url: /static/puzzles/7-mysi_a_syr.pdf
+        level: '2'
+      - node:
+          id: '539563487'
+        name: puzzles-4
+        type: puzzles
+        description: šifry sady 2
+        url: /static/puzzles/7-mysi_a_syr.pdf
+        level: '2'
+      - node:
+          id: '539563487'
+        name: puzzles-5
+        type: puzzles
+        description: šifry sady 2
+        url: /static/puzzles/7-mysi_a_syr.pdf
+        level: '2'
+      - node:
+        id: '539563487'
+        name: checkpoint
+        type: checkpoint
+        description: Kontrola
+        url: ''
+        level: '0'
+      - node:
+          id: '600897437'
+        name: badge-1-1
+        type: badge
+        description: Od značek 11
+        url: badge11.png
+        level: '1'
+      - node:
+          id: '600897447'
+        name: badge-1-2
+        type: badge
+        description: Od značek 12
+        url: badge12.png
+        level: '1'
+      - node:
+          id: '298097758'
+        name: badge-1-3
+        type: badge
+        description: Od značek 13
+        url: badge13.png
+        level: '1'
+      - node:
+          id: '298097759'
+        name: badge-1-4
+        type: badge
+        description: Od značek 13
+        url: badge14.png
+        level: '1'
+      - node:
+          id: '600897437'
+        name: badge-2-1
+        type: badge
+        description: Od značek 21
+        url: badge21.png
+        level: '2'
+      - node:
+          id: '600897447'
+        name: badge-2-2
+        type: badge
+        description: Od značek 22
+        url: badge22.png
+        level: '2'
+      - node:
+          id: '298097758'
+        name: badge-2-3
+        type: badge
+        description: Od značek 23
+        url: badge23.png
+        level: '2'
+      - node:
+          id: '600897437'
+        name: badge-3-1
+        type: badge
+        description: Od značek 31
+        url: badge31.png
+        level: '3'
+      - node:
+          id: '600897447'
+        name: badge-3-2
+        type: badge
+        description: Od značek 32
+        url: badge32.png
+        level: '3'
+      - node:
+          id: '600897437'
+        name: badge-4-1
+        type: badge
+        description: Od značek 41
+        url: badge41.png
+        level: '4'
+      - node:
+          id: '600897437'
+        name: badge-5-1
+        type: badge
+        description: Od značek 51
+        url: badge51.png
+        level: '5'
+  bonuses:
+    bonus:
+      - label: bonus-1
+        description: bonus 1
+        url: ''
+        display-time: '2020-11-06T21:00:00+01:00'
+      - label: bonus-2
+        description: bonus 2
+        url: ''
+        display-time: '2020-11-06T22:00:00+01:00'
+      - label: bonus-3
+        description: bonus 3
+        url: ''
+        display-time: '2020-11-06T23:00:00+01:00'
+      - label: bonus-4
+        description: bonus 4
+        url: ''
+        display-time: '2020-11-07T00:00:00+01:00'
+      - label: bonus-5
+        description: bonus 5
+        url: ''
+        display-time: '2020-11-07T01:00:00+01:00'
+      - label: bonus-6
+        description: bonus 6
+        url: ''
+        display-time: '2020-11-07T02:00:00+01:00'
+      - label: bonus-7
+        description: bonus 7
+        url: ''
+        display-time: '2020-11-07T03:00:00+01:00'
+      - label: bonus-8
+        description: bonus 8
+        url: ''
+        display-time: '2020-11-07T04:00:00+01:00'
+      - label: bonus-9
+        description: bonus 9
+        url: ''
+        display-time: '2020-11-07T05:00:00+01:00'
+      - label: bonus-10
+        description: bonus 10
+        url: ''
+        display-time: '2020-11-07T06:00:00+01:00'
+  name: Kvalifikace Tmou XXII
+
